requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain-openai==0.1.1
langchain-astradb==0.3.3
langchain-core==0.1.52
pandas==2.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
astrapy==1.0.0